Added All the user Case
[18-Jan-16 11:51:06 PM] Balaji Boominathan: 1) User opens the app,
current location is detected, a circle is drawn, data is loaded in the
list according to the categories and map markers are placed. When
Zoom-in no changes in the data or circle, just zoom-in to see the exact
streets and roads. When Zoom-out auto load the data and map markers only
when the center point of the map and last point is atleast 2 km
difference. (just to avoid too much load in the database).
[18-Jan-16 11:52:10 PM] Balaji Boominathan: 2) When the user the user
drags it or Zoomout is more than 2km radius from the last location, do
the steps as in UseCase1.
[18-Jan-16 11:52:34 PM] Balaji Boominathan: 3) By pressing the location
button, user should come to UseCase1.
4) When User search any location in the map it will show all the
location based on that location within 2KM